refactor: accept protocol entry
